refactor: Reorganiza a estrutura de arquivos em diretórios src e data
