+++ data/letter_map.json
@@ -1,0 +1,28 @@
+{
+    "a": ["a", "á", "ah"],
+    "b": ["bê", "be", "b"],
+    "c": ["cê", "ce", "c"],
+    "d": ["dê", "de", "d"],
+    "e": ["e", "é", "eh"],
+    "f": ["efe", "éfe", "f"],
+    "g": ["gê", "ge", "g"],
+    "h": ["agá", "h"],
+    "i": ["i", "í"],
+    "j": ["jota", "j"],
+    "k": ["cá", "ka", "k"],
+    "l": ["ele", "éle", "l"],
+    "m": ["eme", "éme", "m", "em"],
+    "n": ["ene", "éne", "n", "en"],
+    "o": ["o", "ó", "oh"],
+    "p": ["pê", "pe", "p"],
+    "q": ["quê", "que", "q"],
+    "r": ["erre", "érre", "r"],
+    "s": ["esse", "ésse", "s", "es", "és"],
+    "t": ["tê", "te", "t"],
+    "u": ["u"],
+    "v": ["vê", "ve", "v"],
+    "w": ["dáblio", "dablio", "w"],
+    "x": ["xis", "x", "chis"],
+    "y": ["ípsilon", "ipsilon", "y"],
+    "z": ["zê", "ze", "z"]
+}
