+++ data/words.txt
@@ -1,0 +1,42 @@
+casa
+carro
+mar
+ar
+agua
+terra
+povo
+fogo
+mãe
+pai
+filho
+cota
+palhaço
+pássaro
+abelha
+cachorro
+girassol
+elefante
+galinha
+criança
+caminhão
+tesoura
+xícara
+relógio
+foguete
+aquário
+helicóptero
+bicicleta
+floresta
+vermelho
+amarelo
+futebol
+mágico
+árvore
+ônibus
+maçã
+televisão
+brinquedo
+esquilo
+homem
+nariz
+queijo
