+++ requirements.txt
@@ -1,2 +1,4 @@
 customtkinter
 qi @ https://files.pythonhosted.org/packages/da/23/01999b202e1ada1e98460935f52442952a080921bcd7b10d078f95a7bc91/qi-3.1.5-cp311-cp311-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
+thefuzz
+python-Levenshtein
