--- words.txt
@@ -1,12 +1,0 @@
-casa
-carro
-mar
-ar
-agua
-terra
-povo
-fogo
-mãe
-pai
-filho
-cota
